pipeline/requirements.txt
requests==2.32.4
opencv-python==4.12.0.88
numpy==2.2.6
pillow==11.3.0
matplotlib==3.10.3
folium==0.20.0
pandas==2.3.1
pytz==2025.2
transformers==4.54.1
torch==2.7.1
torchvision==0.22.1
scikit-learn==1.7.1
joblib==1.5.1